Refactoring and added exclude feature, missing unit tests
