+++ main.go
@@ -7,31 +7,29 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"retailer/cmd"
 )
 
 type excludes []string
 
+var excludedFiles excludes
+
 func main() {
 
-	var excludedFiles excludes
-
 	maxTails := flag.Int("m", 10, "Max tails running at once")
-	flag.Var(&excludedFiles, "exclude","File to be excluded")
+	flag.Var(&excludedFiles, "exclude", "File to be excluded")
 	flag.Parse()
 	if len(flag.Args()) < 1 {
 		println("Must provide path to monitor")
 		os.Exit(1)
 	}
-fmt.Println("Excludes: ", excludedFiles)
-	cmd.MonitorDir(flag.Arg(0), *maxTails)
+	MonitorDir(flag.Arg(0), *maxTails)
 }
 
 func (n *excludes) Set(value string) error {
-    *n = append(*n, value)
-    return nil
+	*n = append(*n, value)
+	return nil
 }
 
 func (n *excludes) String() string {
-    return fmt.Sprintf("%s", *n)
+	return fmt.Sprintf("%s", *n)
 }
